refactor: split compiler.py into vyper/compiler subpackage
